# docs/swagger.yaml
definitions:
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_advantage.Advantage:
    properties:
      created_at:
        type: string
      id:
        type: string
      image:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_auth.Login:
    properties:
      email:
        type: string
      isHashed:
        type: boolean
      password:
        type: string
    required:
    - email
    - isHashed
    - password
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_auth.Registration:
    properties:
      email:
        type: string
      isHashed:
        type: boolean
      name:
        type: string
      password:
        type: string
    required:
    - email
    - isHashed
    - name
    - password
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /advantages/{advantageId}:
    delete:
      consumes:
      - application/json
      description: Delete advantage by id
      parameters:
      - description: advantage id
        in: path
        name: advantageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Delete Advantage
      tags:
      - advantages
    patch:
      consumes:
      - application/json
      description: Update advantage by id
      parameters:
      - description: advantage id
        in: path
        name: advantageId
        required: true
        type: string
      - description: image file
        in: formData
        name: image
        type: file
      - description: title of advantage
        in: formData
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_advantage.Advantage'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Update Advantage
      tags:
      - advantages
  /advantages/all:
    get:
      consumes:
      - application/json
      description: Get all advantages
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_advantage.Advantage'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get advantages
      tags:
      - advantages
  /advantages/create:
    post:
      consumes:
      - multipart/form-data
      description: Create advantage
      parameters:
      - description: image file
        in: formData
        name: image
        required: true
        type: file
      - description: title of advantage
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Create Advantage
      tags:
      - advantages
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates an existing user and returns a JWT token
      parameters:
      - description: Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_auth.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Login an existing user
      tags:
      - auth
...
  /registration:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_auth.Registration'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"